More changes on the zetteldesk, an emacs package I am working on
I will probably make a seperate repository for it, when its in a more
shareable state, but for now, I just manage it from this repo. That
will continue to happen, as that library is and will be a part of my
dotfiles, but soon there will also be a more dedicated place for it.